форма добавления точки маршрута с моковыми данными
