+++ src/view/trip-route-edit-point.js
@@ -1,41 +1,7 @@
 import dayjs from 'dayjs';
-import {TYPES, DESTINATIONS, OFFER_LIST} from '../const.js';
-
-/**
- * Функция создания разметки дополнительный опций маршрута
- * @param {array} currentOffers - массив объектов, поедставляющий собой выбранные дополнительные опции маршрута
- * @returns - строка, содержащая разметку дополнительных опций маршрута
- */
-const createTripRouteOffer = (currentOffers) => {
-  return OFFER_LIST.map((offer) => {
-    const isChecked = (currentOffers.find((currentOffer) => currentOffer.name === offer.name))? 'checked':'';
-    return `<div class="event__offer-selector">
-    <input class="event__offer-checkbox  visually-hidden" id="event-offer-${offer.type}-1" type="checkbox" name="event-offer-${offer.type}" ${isChecked}>
-    <label class="event__offer-label" for="event-offer-${offer.type}-1">
-      <span class="event__offer-title">${offer.name}</span>
-      &plus;&euro;&nbsp;
-      <span class="event__offer-price">${offer.price}</span>
-    </label>
-  </div>`;
-  }).join('');
-};
-
-/**
- * Функция создания разметки для выпадающего меню типов маршрутов
- * @param {string} currentType - текущий тип маршрута
- * @returns - строка, содержащая разметку для выпадающего меню типов маршрутов
- */
-const createTripRouteEditTypesTemplate = (currentType) => {
-  return TYPES.map((type) => {
-    const isChecked = (type === currentType)? 'checked':'';
-    type = type.toLowerCase();
-    return `<div class="event__type-item">
-      <input id="event-type-${type}-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${type}" ${isChecked}>
-      <label class="event__type-label  event__type-label--${type}" for="event-type-${type}-1">${type}</label>
-    </div>`;
-  }).join('');
-};
-
+import {DESTINATIONS} from '../const.js';
+import {createTripRouteTypesTemplate} from './trip-route-types.js';
+import {createTripRouteOfferTemplate} from './trip-route-offer.js';
 /**
  * Функция создания блока разметки для блока редактирования точки маршрута
  * @param {object} point - объект с данными о точке маршрута
@@ -57,7 +23,7 @@ export const createTripRouteEditPointTemplate = (point) => {
         <div class="event__type-list">
           <fieldset class="event__type-group">
             <legend class="visually-hidden">Event type</legend>
-            ${createTripRouteEditTypesTemplate(type)}
+            ${createTripRouteTypesTemplate(type)}
           </fieldset>
         </div>
       </div>
@@ -99,7 +65,7 @@ export const createTripRouteEditPointTemplate = (point) => {
         <h3 class="event__section-title  event__section-title--offers">Offers</h3>
 
         <div class="event__available-offers">
-          ${createTripRouteOffer(offers)}
+          ${createTripRouteOfferTemplate(offers)}
         </div>
       </section>
